requirements.txt
psycopg==3.2.4
Django==5.1.5
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
boto3==1.36.17
coverage==7.6.11
moto==5.0.28
pillow==11.1.0
requests==2.32.3
django_redis==5.4.0
redis==5.2.1
grpcio==1.70.0
grpcio-tools==1.70.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
typing==3.7.4.3